+++ spring-boot-project/spring-boot-actuator-autoconfigure/src/main/java/org/springframework/boot/actuate/autoconfigure/logging/otlp/OtlpLoggingProperties.java
@@ -50,7 +50,7 @@ public class OtlpLoggingProperties {
 	private Duration connectTimeout = Duration.ofSeconds(10);
 
 	/**
-	 * Transport used to send the spans.
+	 * Transport used to send the logs.
 	 */
 	private Transport transport = Transport.HTTP;
