+++ spring-boot-project/spring-boot-docker-compose/src/main/java/org/springframework/boot/docker/compose/service/connection/otlp/OpenTelemetryTracingDockerComposeConnectionDetailsFactory.java
@@ -56,19 +56,19 @@ class OpenTelemetryTracingDockerComposeConnectionDetailsFactory
 
 		private final int grpcPort;
 
-		private final int httPort;
+		private final int httpPort;
 
 		private OpenTelemetryTracingDockerComposeConnectionDetails(RunningService source) {
 			super(source);
 			this.host = source.host();
 			this.grpcPort = source.ports().get(OTLP_GRPC_PORT);
-			this.httPort = source.ports().get(OTLP_HTTP_PORT);
+			this.httpPort = source.ports().get(OTLP_HTTP_PORT);
 		}
 
 		@Override
 		public String getUrl(Transport transport) {
 			int port = switch (transport) {
-				case HTTP -> this.httPort;
+				case HTTP -> this.httpPort;
 				case GRPC -> this.grpcPort;
 			};
 			return "http://%s:%d/v1/traces".formatted(this.host, port);
