+++ spring-boot-project/spring-boot-testcontainers/src/dockerTest/java/org/springframework/boot/testcontainers/lifecycle/TestcontainersParallelStartupWithImportTestcontainersIntegrationTests.java
@@ -19,15 +19,14 @@ package org.springframework.boot.testcontainers.lifecycle;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.testcontainers.containers.PostgreSQLContainer;
+import org.testcontainers.junit.jupiter.Container;
 
 import org.springframework.boot.test.system.CapturedOutput;
 import org.springframework.boot.test.system.OutputCaptureExtension;
-import org.springframework.boot.testcontainers.lifecycle.TestContainersParallelStartupIntegrationTests.ContainerConfig;
+import org.springframework.boot.testcontainers.context.ImportTestcontainers;
+import org.springframework.boot.testcontainers.lifecycle.TestcontainersParallelStartupWithImportTestcontainersIntegrationTests.Containers;
 import org.springframework.boot.testsupport.container.DisabledIfDockerUnavailable;
 import org.springframework.boot.testsupport.container.TestImage;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.TestPropertySource;
 import org.springframework.test.context.junit.jupiter.SpringExtension;
 
@@ -39,34 +38,27 @@ import static org.assertj.core.api.Assertions.assertThat;
  * @author Phillip Webb
  */
 @ExtendWith(SpringExtension.class)
-@ContextConfiguration(classes = ContainerConfig.class)
 @TestPropertySource(properties = "spring.testcontainers.beans.startup=parallel")
 @DisabledIfDockerUnavailable
 @ExtendWith(OutputCaptureExtension.class)
-class TestContainersParallelStartupIntegrationTests {
+@ImportTestcontainers(Containers.class)
+class TestcontainersParallelStartupWithImportTestcontainersIntegrationTests {
 
 	@Test
 	void startsInParallel(CapturedOutput out) {
 		assertThat(out).contains("-lifecycle-0").contains("-lifecycle-1").contains("-lifecycle-2");
 	}
 
-	@Configuration(proxyBeanMethods = false)
-	static class ContainerConfig {
+	static class Containers {
 
-		@Bean
-		static PostgreSQLContainer<?> container1() {
-			return TestImage.container(PostgreSQLContainer.class);
-		}
+		@Container
+		static PostgreSQLContainer<?> container1 = TestImage.container(PostgreSQLContainer.class);
 
-		@Bean
-		static PostgreSQLContainer<?> container2() {
-			return TestImage.container(PostgreSQLContainer.class);
-		}
+		@Container
+		static PostgreSQLContainer<?> container2 = TestImage.container(PostgreSQLContainer.class);
 
-		@Bean
-		static PostgreSQLContainer<?> container3() {
-			return TestImage.container(PostgreSQLContainer.class);
-		}
+		@Container
+		static PostgreSQLContainer<?> container3 = TestImage.container(PostgreSQLContainer.class);
 
 	}
